# speed up statsd parser