### Required for all PRs:

- [ ] CHANGELOG.md updated

### Before
```
BenchmarkParse-4                                      50000         35152 ns/op
BenchmarkParseWithTemplate-4                          30000         45953 ns/op
BenchmarkParseWithTemplateAndFilter-4                 30000         50166 ns/op
BenchmarkParseWith2TemplatesAndFilter-4               20000         64277 ns/op
BenchmarkParseWith2Templates3TagsAndFilter-4          20000         81050 ns/op
```

### After
```
BenchmarkParse-4                               	  100000	     20525 ns/op
BenchmarkParseWithTemplate-4                   	  100000	     20438 ns/op
BenchmarkParseWithTemplateAndFilter-4          	  100000	     20692 ns/op
BenchmarkParseWith2TemplatesAndFilter-4        	  100000	     20724 ns/op
BenchmarkParseWith2Templates3TagsAndFilter-4   	  100000	     22354 ns/op
```